Move editor components
