+++ src/components/editor/selectors/color-selector.tsx
